Demo 1 Update
Images files moved to src directory into a package
BattleView is establish, but not perfect
BattleController is set up, but does not interact with View much yet and
hardly uses the Model at all.  Needs to be properly wired together.
App.java has been updated to throw in some test data to use for the next
while